Modularize category routes
