+++ lib/server.js
@@ -2,37 +2,22 @@
 
 const express = require('express');
 
+// Middleware
 const requestTime = require('./middleware/request-time');
 const logger = require('./middleware/logger');
 const errorHandler = require('./middleware/error-handler');
 const validator = require('./middleware/validator');
 
-const app = express();
+// Routes
+const categoryRoutes = require('./routes/category-routes');
 
-let db = [];
+const app = express();
 
 app.use(express.json());
 app.use(requestTime);
 
-// Route to Get All Categories
-app.get('/categories', logger('Get all categories'), (req, res, next) => {
-  let count = db.length;
-  let results = db;
-  res.json({ count, results });
-});
-
-// Route to Create a Category
-app.post('/categories', logger('Create a category'), validator, (req, res, next) => {
-  console.log(req.valid);
-  if (!req.valid) {
-    throw new Error();
-    return;
-  }
-  let record = req.body;
-  record.id = Math.random();
-  db.push(record);
-  res.json(record);
-});
+app.get('/categories', logger('Get all categories'), categoryRoutes.get);
+app.post('/categories', logger('Get all categories'), validator, categoryRoutes.post);
 
 app.use(errorHandler);
